delete and reset block fix

+++ Disintegration/Assets/src/Block.cs
@@ -25,6 +25,10 @@ public class Block : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHan
     Vector2 startPosition;
     private GameObject oldBlockGO;
     public static Transform startParent;
+    bool haveNoNaibours;
+
+    private int oldCellX = 0;
+    private int oldCellY = 0;
 
     public void TableUpdate()
     {
@@ -76,6 +80,39 @@ public class Block : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHan
     {
 
         transform.position = Input.mousePosition + new Vector3(-50f, size * 100f, 0f);
+        Vector2 blockPos = GetComponent<RectTransform>().anchoredPosition;
+        if (blockPos.x > 451f && blockPos.x < 2349f && blockPos.y > (-900f + 100f * size) && blockPos.y < 0f)
+        {
+            int cellX = (int)Mathf.Round(blockPos.x / 100f);
+            int cellY = -(int)Mathf.Round(blockPos.y / 100f);
+
+            haveNoNaibours = true;
+
+            for (int i = 0; i < size; i++)
+            {
+                haveNoNaibours = haveNoNaibours && Table._table[cellY + i, cellX - 1]._block == null;
+                haveNoNaibours = haveNoNaibours && Table._table[cellY + i, cellX + 1]._block == null;
+            }
+
+
+            if ((cellX != oldCellX || cellY != oldCellY) && haveNoNaibours)
+            {
+                for (int i = 0; i < size; i++)
+                {
+                    Table._table[cellY + i, cellX].transform.GetChild(0).gameObject.SetActive(true);
+                    Table._table[oldCellY + i, oldCellX].transform.GetChild(0).gameObject.SetActive(false);
+                }
+                oldCellY = cellY;
+                oldCellX = cellX;
+            }
+        }
+        else
+        {
+            for (int i = 0; i < size; i++)
+            {
+                Table._table[oldCellY + i, oldCellX].transform.GetChild(0).gameObject.SetActive(false);
+            }
+        }
     }
 
     #endregion
@@ -85,7 +122,7 @@ public class Block : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHan
     public void OnEndDrag(PointerEventData eventData)
     {
         Vector2 blockPos = GetComponent<RectTransform>().anchoredPosition;
-        if (blockPos.x > 451f && blockPos.x < 2349f && blockPos.y > (-900f + 100f * size) && blockPos.y < 0f)
+        if (blockPos.x > 451f && blockPos.x < 2349f && blockPos.y > (-900f + 100f * size) && blockPos.y < 0f && haveNoNaibours)
         {
             //sound.Play();
             DeleteOldBlockPositions();
@@ -95,10 +132,30 @@ public class Block : MonoBehaviour, IBeginDragHandler, IDragHandler, IEndDragHan
             row = -cellY;
             GetComponent<RectTransform>().anchoredPosition = new Vector2(100f * cellX, 100f * cellY);
             TableUpdate();
+
+            for (int i = 0; i < size; i++)
+            {
+                Table._table[-cellY + i, cellX].transform.GetChild(0).gameObject.SetActive(false);
+            }
         }
         else
         {
-            GetComponent<RectTransform>().anchoredPosition = startPosition;
+            if (blockPos.y < (-900f + 100f * size))
+            {
+                DeleteOldBlockPositions();
+                _grid = transform.parent.GetComponent<Table>();
+                _grid.Refresh();
+                _button.GetComponent<ButtonFiltre>().increment(); 
+                Destroy(gameObject); 
+            }
+            else
+            {
+                for (int i = 0; i < size; i++)
+                {
+                    Table._table[oldCellY + i, oldCellX].transform.GetChild(0).gameObject.SetActive(false);
+                }
+                GetComponent<RectTransform>().anchoredPosition = startPosition;
+            }
         }
         Destroy(oldBlockGO);
     }
